// src/types/mock-data.type.ts
export type MockData = {
  titles: string[];
  descriptions: string[];
  dates: string[];
  city: string[];
  cityLocation: string[];
  previewImage: string[];
  images: string[];
  isPremium: string[];
  isFavorite: string[];
  rating: string[];
  types: string[];
  bedrooms: string[];
  maxAdults: string[];
  price: string[];
  goods: string[];
  host: string[];
  location: string[];
  cities: string[];
  avatars: string[];
  users: string[];
}